Temporary quirk for PWM generation.
The PWM configuration was wrong, however fixing it would end up to broke things
because the upper layer was tuned with the wrong PWM configuration under its feet.

This commit introduces a temporary quirk for this (i.e. restore the original
pwm configuration while providing the correct alternative via #ifdef. This can
be toggled when fixing also the upper layers).